+++ Source/CesiumRuntime/Private/Cesium3DTileset.cpp
@@ -1447,9 +1447,18 @@ std::vector<FCesiumCamera> ACesium3DTileset::GetEditorCameras() const {
         !pEditorViewportClient->IsPerspective()) {
       continue;
     }
+    auto GetViewRotation = [pEditorViewportClient]() {
+      if (pEditorViewportClient->bUsingOrbitCamera) {
 
+        return (pEditorViewportClient->GetLookAtLocation() -
+                pEditorViewportClient->GetViewLocation())
+            .Rotation();
+      } else {
+        return pEditorViewportClient->GetViewRotation();
+      }
+    };
     const FVector& location = pEditorViewportClient->GetViewLocation();
-    const FRotator& rotation = pEditorViewportClient->GetViewRotation();
+    const FRotator& rotation = GetViewRotation();
     float fov = pEditorViewportClient->ViewFOV;
     FIntPoint offset;
     FIntPoint size;