+++ CHANGES.md
@@ -10,6 +10,7 @@
 ##### Fixes :wrench:
 
 - Cesium for Unreal now only uses Editor viewports for tile selection if they are visible, real-time, and use a perspective projection. Previously, any viewport with a valid size was used, which could lead to tiles being loaded and rendered unnecessarily.
+- Fixed a bug that causes tiles to dissapear when the Editor viewport is in Orbit mode.
 
 ### v1.16.2 - 2022-08-04